+++ app/forms/transfer/create_project_form.rb
@@ -20,8 +20,9 @@ class Transfer::CreateProjectForm < CreateProjectForm
 
   validates :outgoing_trust_sharepoint_link, presence: true, sharepoint_url: true
 
-  validate :check_incoming_trust_and_outgoing_trust
   validate :outgoing_trust_exists, if: -> { outgoing_trust_ukprn.present? }
+
+  validates_with OutgoingIncomingTrustsUkprnValidator
 
   def initialize(params = {})
     @attributes_with_invalid_values = []
@@ -72,10 +73,6 @@ class Transfer::CreateProjectForm < CreateProjectForm
     nil
   rescue TypeError, Date::Error, NegativeValueError
     @attributes_with_invalid_values << :provisional_transfer_date
-  end
-
-  def check_incoming_trust_and_outgoing_trust
-    errors.add(:incoming_trust_ukprn, I18n.t("errors.attributes.incoming_trust_ukprn.ukprns_must_not_match")) if incoming_trust_ukprn == outgoing_trust_ukprn
   end
 
   private def notify_team_leaders(project)