+++ spec/support/academies_api_helpers.rb
@@ -97,4 +97,35 @@ module AcademiesApiHelpers
     allow(test_client).to receive(:get_establishment).with(urn).and_raise(Api::AcademiesApi::Client::UnauthorisedError)
     allow(Api::AcademiesApi::Client).to receive(:new).and_return(test_client)
   end
+
+  def mock_api_for_editing
+    establishment = build(:academies_api_establishment)
+    local_authority = build(:local_authority)
+    allow(establishment).to receive(:local_authority).and_return(local_authority)
+
+    mock_establishment = Api::AcademiesApi::Client::Result.new(establishment, nil)
+
+    trust = build(:academies_api_trust, ukprn: 10059151)
+    mock_before_incoming_trust = Api::AcademiesApi::Client::Result.new(trust, nil)
+
+    trust = build(:academies_api_trust, ukprn: 10058882)
+    mock_after_incoming_trust = Api::AcademiesApi::Client::Result.new(trust, nil)
+
+    trust = build(:academies_api_trust, ukprn: 10059797)
+    mock_before_outgoing_trust = Api::AcademiesApi::Client::Result.new(trust, nil)
+
+    trust = build(:academies_api_trust, ukprn: 10064639)
+    mock_after_outgoing_trust = Api::AcademiesApi::Client::Result.new(trust, nil)
+
+    test_client = Api::AcademiesApi::Client.new
+
+    allow(test_client).to receive(:get_establishment).with(123456).and_return(mock_establishment)
+    allow(test_client).to receive(:get_trust).with(10059151).and_return(mock_before_incoming_trust)
+    allow(test_client).to receive(:get_trust).with(10058882).and_return(mock_after_incoming_trust)
+
+    allow(test_client).to receive(:get_trust).with(10059797).and_return(mock_before_outgoing_trust)
+    allow(test_client).to receive(:get_trust).with(10064639).and_return(mock_after_outgoing_trust)
+
+    allow(Api::AcademiesApi::Client).to receive(:new).and_return(test_client)
+  end
 end